Fix bug

+++ SignalRAssignment_SE151295/SalesRazorPageApp/Controllers/PostsController.cs
@@ -25,6 +25,20 @@ namespace SalesRazorPageApp.Controllers
             _signalRHub = signalRHub;
         }
 
+        public void checkPermission()
+        {
+
+                //ViewBag.IsMemberLogin = "False";
+            
+
+/*                ViewBag.IsMemberLogin = "True";
+                if (userRepository.GetCurrentMemberLogin() != null)
+                {
+                    ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
+                }*/
+            
+        }
+
         // GET: Posts
         public  IActionResult Index()
         {
@@ -33,7 +47,18 @@ namespace SalesRazorPageApp.Controllers
                 return RedirectToAction("ErrorSession", "Login");
             }
 
-            ViewBag.IsMemberLogin = userRepository.CheckIsMemberLogin().ToString();
+            ViewBag.IsMemberLogin = "False";
+            return View();
+        }
+
+        public IActionResult IndexUser()
+        {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToAction("ErrorSession", "Login");
+            }
+
+            ViewBag.IsMemberLogin = "True";
             if (userRepository.GetCurrentMemberLogin() != null)
             {
                 ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
@@ -54,11 +79,7 @@ namespace SalesRazorPageApp.Controllers
         public IActionResult SearchPostView(string search)
         {
             var listSearch = repository.SearchPost(search);
-            ViewBag.IsMemberLogin = userRepository.CheckIsMemberLogin().ToString();
-            if (userRepository.GetCurrentMemberLogin() != null)
-            {
-                ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
-            }
+            checkPermission();
             if (listSearch == null || listSearch.Count == 0)
             {
                 ViewBag.ErrMsgSearch = "Search not found!";
@@ -73,13 +94,36 @@ namespace SalesRazorPageApp.Controllers
             {
                 return RedirectToAction("ErrorSession", "Login");
             }
+            ViewBag.IsMemberLogin = "False";
+
             ViewData["AuthorID"] = new SelectList(repository.GetAppUserList(), "UserID", "FullName");
             ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryName");
             return View();
         }
 
+        // GET: Posts/Create
+        public IActionResult CreateUser()
+        {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToAction("ErrorSession", "Login");
+            }
+
+            ViewBag.IsMemberLogin = "True";
+            if (userRepository.GetCurrentMemberLogin() != null)
+            {
+                ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
+            }
+
+            List<AppUsers> listApp = new List<AppUsers>();
+            listApp.Add(userRepository.GetCurrentMemberLogin());
+            ViewData["AuthorID"] = new SelectList(listApp, "UserID", "FullName");
+            ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryName");
+            return View();
+        }
+
         // POST: Posts/Create
-         [HttpPost]
+        [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PostID,AuthorID,CreatedDate,UpdatedDate,Title,Content,PublishStatus,CategoryID")] Posts posts)
         {
@@ -87,6 +131,9 @@ namespace SalesRazorPageApp.Controllers
             {
                 return RedirectToAction("ErrorSession", "Login");
             }
+            ViewBag.IsMemberLogin = "False";
+
+
             if (ModelState.IsValid)
             {
                 repository.CreatePost(posts);
@@ -95,6 +142,37 @@ namespace SalesRazorPageApp.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["AuthorID"] = new SelectList(repository.GetAppUserList(), "UserID", "FullName", posts.AuthorID);
+            ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryName", posts.CategoryID);
+            return View(posts);
+        }
+
+        // POST: Posts/Create
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> CreateUser([Bind("PostID,AuthorID,CreatedDate,UpdatedDate,Title,Content,PublishStatus,CategoryID")] Posts posts)
+        {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToAction("ErrorSession", "Login");
+            }
+
+            ViewBag.IsMemberLogin = "True";
+            if (userRepository.GetCurrentMemberLogin() != null)
+            {
+                ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
+            }
+
+            if (ModelState.IsValid)
+            {
+                repository.CreatePost(posts);
+                TempData["SuccessMessage"] = "Create successfully";
+                await _signalRHub.Clients.All.SendAsync("LoadPosts");
+                return RedirectToAction(nameof(IndexUser));
+            }
+
+            List<AppUsers> listApp = new List<AppUsers>();
+            listApp.Add(userRepository.GetCurrentMemberLogin());
+            ViewData["AuthorID"] = new SelectList(listApp, "UserID", "FullName");
             ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryName", posts.CategoryID);
             return View(posts);
         }
@@ -111,6 +189,37 @@ namespace SalesRazorPageApp.Controllers
             {
                 return NotFound();
             }
+            ViewBag.IsMemberLogin = "False";
+
+
+            var posts = repository.GetPostById((int)id);
+            if (posts == null)
+            {
+                return NotFound();
+            }
+            ViewData["AuthorID"] = new SelectList(repository.GetAppUserList(), "UserID", "UserID", posts.AuthorID);
+            ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryName", posts.CategoryID);
+            return View(posts);
+        }
+
+        // GET: Posts/Edit/5
+        public IActionResult EditUser(int? id)
+        {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToAction("ErrorSession", "Login");
+            }
+
+            if (id == null)
+            {
+                return NotFound();
+            }
+            ViewBag.IsMemberLogin = "True";
+            if (userRepository.GetCurrentMemberLogin() != null)
+            {
+                ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
+            }
+            
 
             var posts = repository.GetPostById((int)id);
             if (posts == null)
@@ -123,7 +232,7 @@ namespace SalesRazorPageApp.Controllers
         }
 
         // POST: Posts/Edit/5
-       [HttpPost]
+        [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int PostID, [Bind("PostID,AuthorID,CreatedDate,UpdatedDate,Title,Content,PublishStatus,CategoryID")] Posts posts)
         {
@@ -135,6 +244,8 @@ namespace SalesRazorPageApp.Controllers
             {
                 return NotFound();
             }
+
+            ViewBag.IsMemberLogin = "False";
 
             if (ModelState.IsValid)
             {
@@ -150,6 +261,46 @@ namespace SalesRazorPageApp.Controllers
                 }
 
                 return RedirectToAction(nameof(Index));
+            }
+            ViewData["AuthorID"] = new SelectList(repository.GetAppUserList(), "UserID", "UserID", posts.AuthorID);
+            ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryID", posts.CategoryID);
+            return View(posts);
+        }
+
+        // POST: Posts/Edit/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> EditUser(int PostID, [Bind("PostID,AuthorID,CreatedDate,UpdatedDate,Title,Content,PublishStatus,CategoryID")] Posts posts)
+        {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToAction("ErrorSession", "Login");
+            }
+            if (PostID != posts.PostID)
+            {
+                return NotFound();
+            }
+
+            ViewBag.IsMemberLogin = "True";
+            if (userRepository.GetCurrentMemberLogin() != null)
+            {
+                ViewBag.UserName = userRepository.GetCurrentMemberLogin().FullName;
+            }
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    TempData["SuccessMessage"] = "Edit successfully";
+                    repository.UpdatePost(posts);
+                    await _signalRHub.Clients.All.SendAsync("LoadPosts");
+                }
+                catch
+                {
+
+                }
+
+                return View(nameof(IndexUser));
             }
             ViewData["AuthorID"] = new SelectList(repository.GetAppUserList(), "UserID", "UserID", posts.AuthorID);
             ViewData["CategoryID"] = new SelectList(repository.GetCategoryList(), "CategoryID", "CategoryID", posts.CategoryID);
@@ -188,6 +339,8 @@ namespace SalesRazorPageApp.Controllers
             {
                 return RedirectToAction("ErrorSession", "Login");
             }
+            checkPermission();
+
 
             var posts = repository.GetPostById(PostID);
             repository.DeletePost(posts);
